webapp/requirements.txt
Flask==3.1.0
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
numpy==2.2.2
pandas==2.2.3
Werkzeug==3.1.3
python-dotenv
google-search-results
requests==2.31.0
Faker==19.13.0
psycopg2-binary
Flask-Migrate